+++ tests/test_querysetsequence.py
@@ -119,6 +119,11 @@ class TestLength(TestBase):
         with self.assertNumQueries(2):
             self.assertEqual(self.all.count(), 5)
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_acount(self):
+        # The proper length should be returned via database queries.
+        self.assertEqual(await self.all.acount(), 5)
+
     def test_len(self):
         # Calling len() evaluates the QuerySet.
         with self.assertNumQueries(2):
@@ -138,9 +143,19 @@ class TestLength(TestBase):
         with self.assertNumQueries(4):
             self.assertEqual(len(self.all[1:]), 4)
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_acount_slice(self):
+        """Ensure the proper length is calculated when a slice is taken."""
+        self.assertEqual(await self.all[1:].acount(), 4)
+
     def test_empty_count(self):
         """An empty QuerySetSequence has a count of 0."""
         self.assertEqual(self.empty.count(), 0)
+
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_acount_empty_count(self):
+        """An empty QuerySetSequence has a count of 0."""
+        self.assertEqual(await self.empty.acount(), 0)
 
     def test_empty_len(self):
         """An empty QuerySetSequence has a count of 0."""
@@ -1380,11 +1395,26 @@ class TestGet(TestBase):
         self.assertEqual(book.title, "Biography")
         self.assertIsInstance(book, Book)
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aget(self):
+        """
+        Ensure that aget() returns the expected element or raises DoesNotExist.
+        """
+        # Get a particular item.
+        book = await self.all.aget(title="Biography")
+        self.assertEqual(book.title, "Biography")
+        self.assertIsInstance(book, Book)
+
     def test_not_found(self):
         # An exception is raised if get() is called and nothing is found.
         with self.assertNumQueries(2):
             with self.assertRaises(ObjectDoesNotExist):
                 self.all.get(title="")
+
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aget_not_found(self):
+        with self.assertRaises(ObjectDoesNotExist):
+            await self.all.aget(title="")
 
     def test_multi_found(self):
         """Test multiple found in the same QuerySet."""
@@ -1393,6 +1423,14 @@ class TestGet(TestBase):
             with self.assertRaises(MultipleObjectsReturned):
                 self.all.get(author=self.bob)
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aget_multi_found(self):
+        """Test multiple found in the same QuerySet."""
+        # ...or if aget() is called and multiple objects are found.
+        with self.assertRaises(MultipleObjectsReturned):
+            await self.all.aget(author=self.bob)
+
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
     def test_multi_found_separate_querysets(self):
         """Test one found in each QuerySet."""
         # ...or if get() is called and multiple objects are found.
@@ -1400,10 +1438,24 @@ class TestGet(TestBase):
             with self.assertRaises(MultipleObjectsReturned):
                 self.all.get(title__contains="A")
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aget_multi_found_separate_querysets(self):
+        """Test one found in each QuerySet."""
+        # ...or if aget() is called and multiple objects are found.
+        with self.assertRaises(MultipleObjectsReturned):
+            await self.all.aget(title__contains="A")
+
     def test_related_model(self):
         qss = QuerySetSequence(Article.objects.all(), BlogPost.objects.all())
         with self.assertNumQueries(2):
             post = qss.get(publisher__name="Wacky Website")
+        self.assertEqual(post.title, "Post")
+        self.assertIsInstance(post, BlogPost)
+
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aget_related_model(self):
+        qss = QuerySetSequence(Article.objects.all(), BlogPost.objects.all())
+        post = await qss.aget(publisher__name="Wacky Website")
         self.assertEqual(post.title, "Post")
         self.assertIsInstance(post, BlogPost)
 
@@ -1414,10 +1466,23 @@ class TestGet(TestBase):
         self.assertEqual(article.title, "Some Article")
         self.assertIsInstance(article, Article)
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aget_queryset_lookup(self):
+        """Test using the special QuerySet lookup."""
+        article = await self.all.aget(**{"#": 1, "author": self.bob})
+        self.assertEqual(article.title, "Some Article")
+        self.assertIsInstance(article, Article)
+
     def test_empty(self):
         """Calling get on an empty QuerySetSequence raises ObjectDoesNotExist."""
         with self.assertRaises(ObjectDoesNotExist):
             self.empty.get(pk=1)
+
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aget_empty(self):
+        """Calling get on an empty QuerySetSequence raises ObjectDoesNotExist."""
+        with self.assertRaises(ObjectDoesNotExist):
+            await self.empty.get(pk=1)
 
 
 class TestBoolean(TestBase):
@@ -1553,6 +1618,13 @@ class TestExists(TestBase):
         with self.assertNumQueries(1):
             self.assertTrue(self.all.filter(title="Biography").exists())
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aexists(self):
+        """
+        Ensure that aexists() returns True if the item is found in the first QuerySet.
+        """
+        self.assertTrue(await self.all.filter(title="Biography").aexists())
+
     def test_exists_second(self):
         """
         Ensure that exists() returns True if the item is found in a subsequent QuerySet.
@@ -1560,19 +1632,42 @@ class TestExists(TestBase):
         with self.assertNumQueries(2):
             self.assertTrue(self.all.filter(title="Alice in Django-land").exists())
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aexists_second(self):
+        """
+        Ensure that aexists() returns True if the item is found in a subsequent
+        QuerySet.
+        """
+        self.assertTrue(await self.all.filter(title="Alice in Django-land").aexists())
+
     def test_not_found(self):
         """Ensure that exists() returns False if the item is not found."""
         with self.assertNumQueries(2):
             self.assertFalse(self.all.filter(title="").exists())
+
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aexists_not_found(self):
+        """Ensure that aexists() returns False if the item is not found."""
+        self.assertFalse(await self.all.filter(title="").aexists())
 
     def test_multi_found(self):
         """Ensure that exists() returns True if multiple items are found."""
         with self.assertNumQueries(1):
             self.assertTrue(self.all.filter(author=self.bob).exists())
 
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aexists_multi_found(self):
+        """Ensure that aexists() returns True if multiple items are found."""
+        self.assertTrue(await self.all.filter(author=self.bob).aexists())
+
     def test_empty(self):
         """An empty QuerySetSequence should return False."""
         self.assertFalse(self.empty.exists())
+
+    @skipIf(django.VERSION < (4, 1), "Not supported in Django < 4.1.")
+    async def test_aexists_empty(self):
+        """An empty QuerySetSequence should return False."""
+        self.assertFalse(await self.empty.aexists())
 
 
 @skipIf(django.VERSION < (4, 0), "Not supported in Django < 4.0.")
@@ -1786,12 +1881,14 @@ class TestNotImplemented(TestCase):
         with self.assertRaises(NotImplementedError):
             self.all.raw("")
 
+    @skipIf(django.VERSION >= (4, 1), "aget exists starting on Django 4.1.")
     async def test_aget(self):
-        with self.assertRaises(ImplementedIn41):
+        with self.assertRaises(AttributeError):
             await self.all.aget()
 
+    @skipIf(django.VERSION >= (4, 1), "acount exists starting on Django 4.1.")
     async def test_acount(self):
-        with self.assertRaises(ImplementedIn41):
+        with self.assertRaises(AttributeError):
             await self.all.acount()
 
     async def test_aiterator(self):
@@ -1822,8 +1919,9 @@ class TestNotImplemented(TestCase):
         with self.assertRaises(ImplementedIn41):
             await self.all.aaggregate()
 
+    @skipIf(django.VERSION >= (4, 1), "aexists exists starting on Django 4.1.")
     async def test_aexists(self):
-        with self.assertRaises(ImplementedIn41):
+        with self.assertRaises(AttributeError):
             await self.all.aexists()
 
     async def test_acontains(self):