+++ CHANGELOG.rst
@@ -16,7 +16,8 @@ Bugfixes
 Improvements
 ------------
 
-* Initial support for `asynchronous queries`_. (`#99 <https://github.com/clokep/django-querysetsequence/pull/99>`_)
+* Initial support for `asynchronous queries`_. (`#99 <https://github.com/clokep/django-querysetsequence/pull/99>`_,
+  `#103 <https://github.com/clokep/django-querysetsequence/pull/103>`_)
 
 .. _asynchronous queries: https://docs.djangoproject.com/en/4.1/topics/db/queries/#async-queries